requirements.txt
django_cors_headers==2.4.0
python_decouple==3.1
stripe==2.10.0
Django==1.11.15
whitenoise==2.0.6
